packaging:Added version manager script
Added helper script to generate changelogs and version strings, taking
into account the special commit message headers:

    Sem-Ver: feature

To bump the middle version and:

    Sem-Ver: .*break.*

To bump the major version.

This also allows avoiding to build wrongly versioned packages from
gerrit, in the period between the commit is submitted and the tag is
sent.

Change-Id: Iedf2fa94cabccd2a2f4071581a3c273d882f73f6
Signed-off-by: David Caro <dcaroest@redhat.com>