libvirt: add auth
Previously, lago could not work with libvirt that required auth. This
can be problematic for people working with projects that do use
authentication in libvirt, one example being VDMS.

We therefore allow setting of libvirt username and password in
lago.conf, and use these values to authenticate. Default values are
empty, therefore not expecting any form of authentication.