Removed redundant checks when getting ovirt api
wait parameter should had to be set to true
in order for any code to be executed in '_get_api',
so it is not needed.

Removed engine's service check in order to
support a situation when we can't check for the
service stauts, but https connection is available.

I think that there is a sense in trying to request
the api within a time window, instead of giving up
on first failure, So I kept the request inside
'assert_true_within_short'.

I had to remove the 'or' condition, otherwise
if '_create_api()' returns false, the condition will
be evaluated to true, which in turn will break the
loop of 'assert_true_within'. Eventually we will end
up without api object.

Signed-off-by: gbenhaim <galbh2@gmail.com>