virtio-scsi support in Lago
1. Removed the USB controller - and added virtio-SCSI controller
with a single dedicated IO thread assigned to it.
2. Removed the bootmenu - did not see a need for it.
3. If a disk begins with sd (for example, 'sda') - assume it's SCSI.
4. Unconditionally add 'discard' support. It works for IDE, virtio-scsi.
It works (though does nothing) with virtio-blk.

Tested by converting some ovirt-system-tests to use sdX drives.
By running 'fstrim' inside the guests saw a good reduction in size of disks.

Note: the root disk cannot be virtio-scsi - requires a change in the lago-image.