Change to repoman options
This change allows repoman to collect
multiple versions of the same package into
the internal repo of the prefix.

Currently packages are being taken to the
internal repo in the following order:

1. sources that were specified using '--custom-source'.
2. Packages from the repos that were specified in the
   reposync-config file (passed to lago with --reposync-yum-config).
   Note: currently, the order in the config matters, for example:
   only package 'x.n' from the first repo will be added to the internal
   repo, even if package 'x.n+1' is in the second repo.

This situation causes unwanted behaviour when there are multiple
versions of the same package in different repos, and this change
is meant to fix it.

Signed-off-by: gbenhaim <galbh2@gmail.com>