More robust CPU selection
CPU model for Lago VMs is passed through from the host and the cluster CPU
family is set based on that CPU model.  With this approach, the environment
deployment may fail due to invalid CPU family of the cluster.  Although the
family corresponds to the host CPU, kernel may remove some CPU flags in nested
virtual machines (when the Lago host is actually a VM).  Then some CPU flags
required for the given CPU family (such as Haswell-NoTSX) may be missing in the
Lago virtual machines.  This is a real problem observed on Lenovo W541.

This patch fixes the problem by selecting the minimum reasonable CPU model.
Westmere is selected unless older or non-Intel hardware is used.  Additionally,
virtualization flags may still be missing with some CPU models (real problem
observed on Lenovo W541 with Westmere).  So let's them to the CPU features,
simply for both Intel and AMD families as optional features, libvirt will do
the right thing with this setting.

A future patch could make the CPU family configurable, when one wants to run
the tests on a different CPU model.