utils: add uuid to LogTask
This adds a new attribute to a log_task, a 'uuid',
if not passed(and thus re-used) the uuid will be generated.

The uuid will appear on all log records that aren't in 'info'
level. It is up to the caller to decide what to do with the  uuid.

Example with debug level:

with LogTask('some_cmd', level='debug'):
	...

lago.log:
.... ::DEBUG::start task:30b10b0a-cb03-4378-b54b-f32ac200d175:some cmd
...
.... ::DEBUG::end task:30b10b0a-cb03-4378-b54b-f32ac200d175:some cmd

And with custom usage:

with LogTask('some cmd') as task:
  LOGGER.debug('my uuid: %s', str(task.uuid))

lago.log:
... ::INFO::@some cmd:
... ::DEBUG::my uuid: 1a6884d5-851d-414d-8b18-53912f9ebd86
... ::INFO::@ some cmd: Success (in 0:00:00)

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>