utils: add to run_command a uuid identifier
This patch will generate a uuid and log it on each call to
'_run_command' method. It also allows to reuse an existing uuid.
This is useful in debugging long-running commands(such as repoman),
especially with the combination of hitting 'shift+*' while the cursor
is standing on the uuid(in VIM at least).

Example:
	run_command(['ls', '-lsah'])

lago.log:
... ::DEBUG::start task:c62d45f1-bae6-40d7-95fa-277b0a5d6305:Run command: "ls" "-lsah":
... ::DEBUG::c62d45f1-bae6-40d7-95fa-277b0a5d6305: command exit with return code: 0
... ::DEBUG::c62d45f1-bae6-40d7-95fa-277b0a5d6305: command stdout: total 44K
4.0K drwxrwxr-x.  3 ngoldin ngoldin 4.0K Jan 22 18:57 .

...

... ::DEBUG::end task:c62d45f1-bae6-40d7-95fa-277b0a5d6305:Run command: "ls" "-lsah":

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>