[WIP] Adding Lago repo plugin
Lago repo plugin should be used for configuring the repo
config on guest vms.

The idea is to take the existing config on the guest, modyfing
it and then writing back (except for when injecting a new config).

RepoPlugin is an object that warps a vm object.

Classes which derived from RepoPlugin should contain the specific
implementaion, according to the distro which installed on the guest.

repo_utils includes helper functions for working with repo configs.

Open Tasks:
============

1. Add tests.
2. Add more exception handling.
3. Add more documentation.
4. Configure all the vms in parallel.
5. Export hard-coded values to the config file.

Signed-off-by: gbenhaim <galbh2@gmail.com>