ovirtlago: adding waiters.py
This new class creates a common infrastructure for tasks that require
polling results periodically from the SDK.
The methods in the 'waiters' class are generated from the data/waiters.yaml
file. Each method defined in the YAML is in charge of polling the Engine
for a status of a request. In this commit only 3 methods were
configured in the YAML:

waiters.vm_up(api, id)
waiters.vm_down(api, id)
waiters.host_up(api, id)

Invoking each of these methods will check if the VM with the given id
reached an 'accepting' state as defined in the YAML 'acceptors' section.
It will run for 'max_attempts', with a 'delay' between and block until.

The polling might end earlier, if the VM reached any of the 'rejectors'
states as defined in the YAML, and then a 'WaiterError' exception will
be thrown.

Adding more 'waiters' requires only updating the YAML.

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>