Stop using sys.exc_info in LogTask
1. sys.exc_info returns the last exception that was caught,
this can confuse LogTask and make it print messages on exceptions that
were already handled and don't interupt to the flow of lago,
(for example when waiting for ssh).
Insted we will use the default arguments which are being passed
to __exit__, and reflects the last exception that was thrown and not handled.

2. If an error occurred, print that stack trace in LogTask.

Signed-off-by: gbenhaim <galbh2@gmail.com>