Control bootstrap in the init file
Adding the ability to specify whether to perform
bootstrap on a given domain or not.

This will be done by adding the 'bootstrap' key and
boolean value to the domain definition.

The default will be to run bootstrap.

Init switch '--skip-bootstrap' has precedence over the
configuration in the init file.

Signed-off-by: gbenhaim <galbh2@gmail.com>