Save the version which was used to generate the prefix
There can be cases where one initialized an environment, and then
upgraded to a newer Lago version. Then, running commands which happens
after the initialization might fail if the new code relays on a newer
metadata saved in the prefix.

This commit adds the Lago version string which was used to generate the
environment under the Prefix metadata, thus providing a way to inspect
which code should be used. It will default to reporting the current
released version(0.36.0) if no version string is found.

Also adding a function to compare versions using pkg_resources.

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>