Extract argparse defaults to constants
This ensures the defaults can be shared both by the CLI and the SDK.
ConfigLoad would get an additional argument of defaults dict, which will
be loaded prior to any configuration files.

Also dropping the '--defaults-only' options from lago generate-config,
as its benefit is little, but supporting it makes the configuration loading
more complex. This means that if you run:

    lago --logdepth=100 generate-config

You will see ... logdepth = 100 in the output, and you can't ignore it.
Which is the expected behaviour I would say(frankly, I don't remember
why I added that option).

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>