sysprep: support debian network scripts locations
In debian, those are found in:
  /etc/network/interfaces
And per-interfaces configurations can be loaded from:
  /etc/network/interfaces.d/

In ubuntu 16.04, the naming scheme changed, (eth -> ens), however, it
can still be configured by updating grub, which is what we'll do for now
when building the images. So keeping 'eth' convention for debian.

Didn't do any major refactoring here to sysprep.py, though we sure need to at
some point.

Note that this does not allow NetworkManager to configure the
interfaces, however, Ubuntu 16.04 does not come with it pre-installed with it,
and installing it adds takes around 250MB.

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>