sysprep: ensure el based ifcfg files are created consistently
Using the SDK I see random failures on boot to acquire IP via DHCP,
looking at the logs, it looks like the order of keys and values in the
generated ifcfg-eth* devices is different than successful runs in the
CLI. Also I see failures when the 'TYPE' parameter is first, while
successes when 'HWADDR' is first. This in all might be vodoo tales, but
it might solve this issue, before digging deeper into what is
wrong with NetworkManager(the logs don't reveal much currently - just
that dhclient start, but it doesn't send any requests).

If this is only vodoo indeed, this patch can't harm(just ensures we use
the same order).

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>