pip: drop 'python-magic' dependency
Functionality wise, from now on we don't try to guess if an image is
compressed using the libmagic python binding, but assume it is
always compressed.

Packaging wise, we no longer depend on 'python-magic'. The
'python-magic' RPM is not the same as the one available in pip, so all
options of keeping it does not worth the effort:
1. Conditional import and working with two different APIs(if the pip
binding is available, or if the RPM binding is available).
2. Keeping the RPM only - prevents PIP installation.
3. Keeping the PIP only - prevents RPM installation.

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>