tests: use sdist instead of RPM in functional SDK
Let tox trigger the sdist build and setup Lago in the virtual
environment for the functional SDK tests. This means, that we are
totally pip installable, leaving our only system-wise dependencies to be
libvirt and libguestfs!

guestfs python binding cannot be installed from PyPI, therefore it
cannot be easily installed from the sdist(as pip by default does not
respect dependency_links). So here, and as should be in the docs, we
just install the python bindings directly from the http link.

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>