libvirt: use /dev/random on libvirt versions <= 2002001
It appears that on older libvirt versions(such as 1.3 on Ubuntu16.04),
/dev/urandom backend model is not supported. I could not trace in libvirt's
changelog when that was changed, and I hope this is not a different
issue specific to Ubuntu16.04.  However, we know for sure that /dev/urandom
works on libvirt version greater than 2002001. So here we add a simple
condition on that.

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>