export: Safely try to parse the name of the backing image
Due to lago image store naming convention (repo:image:version),
if the backing image is a Lago image, preprocessing to the name
needs to be done (in order to extract only the image name) and then
set it as the backing image of the exported image.

Since not always the backing image will be a Lago images, adding
exception handling to the parsing stage.

Signed-off-by: gbenhaim <galbh2@gmail.com>