Use uuid4 instead of uuid1
Until now the prefix uuid was generated with 'uuid1' command. In cases
where you started two environments with the same network names - the uuid1 usage
caused us to pass the same network name to libvirt, despite the environments being
different, thus throwing a libvirt error.

The reason is that we truncated the 8 last bytes of the uuid1 and joined
that with the network name. The 8 last bytes of uuid1 are most
definitely not random.
I see no reason to have the uuid not completely random - therefore this
patch changes the prefix uuid to be generated with uuid4, which should
be completely random.

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>