Catch SIGTERM/SIGHUP in order to run cleanup code
This change will require lago to run cleanup code defined in
"ContextMangers" and "finally" statements even when SIGTERM or
SIGHUP was sent to the process.

This is done by calling "sys.exit" which raises "SystemExit" that
triggers the cleanup.

The exit status will be 128 + the value of the signal, which is
common to Unix programs.

Signed-off-by: gbenhaim <galbh2@gmail.com>