cli-functional-tests: Don't use lago's group
Up until now we used to trigger lago's functional tests with the lago
group.

This can cause an issue if the lago group doesn't exist on the host that
runs the mock. Also, this is not really testing anything, since we run as
root inside the mock.

In order to test the creation of the lago group, tests can be run in
"lago-demo" repository, which uses VMs for running the tests.

Signed-off-by: gbenhaim <galbh2@gmail.com>