DO NOT MERGE: Investigate log collection failure on FC
On FC* log collection functional test fails because of a lock on the VM disk.
This happens because lago fallback to use Libguestfs for collecting
the logs from the vms (and then qemu and Libguestfs tries to access the
VM's disk). During this test, the VM should be available via SSH.

Signed-off-by: gbenhaim <galbh2@gmail.com>