Remove debian based images from functional tests
During the functional tests, the VMs with those images can't
be accessed using SSH. This is caused by a bug in Paramiko, which
cause it to fail parsing the SSH key. It doesn't effect el/fc images,
since those images can be accessed using password authentication.

Until we sort this out, remove those images from the functional tests in
order to not block other PRs.

Signed-off-by: gbenhaim <galbh2@gmail.com>