Ignore deprecation error caused by stevedore
setuptools (which is used by stevedore) 11.3 deprecated the required flag
to load() in favor of two new methods.

A fix was introduced in stevedore 1.2.0.

Since we are using and older version of stevedore (0.14) from epel,
lago outputs the deprecation warning, which isn't helpful for our users.

A long term solution will be to use a newer version of stevedore, but
currently it only available in Openstack repos.

Signed-off-by: gbenhaim <galbh2@gmail.com>