cmd: Fix arg parsing when no verb is provided
With [1] the argument parsing machinery stopped complaining when no verb
is being passed to lago - the code simply falls through with the value
for 'args.verb' equal to 'None'. This patch fixes the issue by
explicitly checking the value of the 'args.verb' argument and
exiting lago when no verb is provided.

[1] https://github.com/python/cpython/commit/f97c59aaba2d93e48cbc6d25f7ff9f9c87f8d0b2