utils: Fix reraising exceptions
'exc_info' usually means a tuple with three elements [1][2]:

- type of the exception
- instance of the exception object
- traceback for the exception

In this particuliar use case, we're collecting exceptions raised
in spawned threads [3] and reraising them in the main thread [4].
The weird thing is that the original code passed the value of the
exception as type... I have no idea how this could possibly work
with py2, but the tests were always fine. With py3 we get expected
errors that we're trying to raise an exception with a value instead
of a type. This patch fixes this issue.

[1] https://docs.python.org/2.7/library/sys.html?highlight=sys%20exc_info#sys.exc_info
[2] https://docs.python.org/3.6/library/sys.html?highlight=exc_info#sys.exc_info
[3] https://github.com/lago-project/lago/blob/9803eeacd41b3f91cd6661a110aa0285aaf4b957/lago/utils.py#L65
[4] https://github.com/lago-project/lago/blob/9803eeacd41b3f91cd6661a110aa0285aaf4b957/lago/utils.py#L97